Move "renderEventItem" engine to new "presenter".
