+++ src/presenter/trip-events-presenter.js
@@ -1,9 +1,8 @@
 import EventsSortView from '../view/events-sort';
 import EventsListView from '../view/events-list';
 import EmptyEventsListView from '../view/empty-events-list';
-import EventItemView from '../view/event-item';
-import EventItemFormView from '../view/event-item-form';
-import {render, RenderPosition, replace} from '../utils/render';
+import {render, RenderPosition} from '../utils/render';
+import EventPresenter from './event-presenter';
 
 export default class TripEventsPresenter {
   constructor(tripEventsContainer) {
@@ -31,38 +30,8 @@ export default class TripEventsPresenter {
   }
 
   _renderEventItem(event) {
-    const eventComponent = new EventItemView(event);
-    const eventEditComponent = new EventItemFormView(event);
-
-    const replaceItemToItemForm = () => {
-      replace(eventEditComponent, eventComponent);
-    };
-
-    const replaceItemFormToItem = () => {
-      replace(eventComponent, eventEditComponent);
-    };
-
-    const onEscKeyDown = (evt) => {
-      if (evt.key === 'Escape' || evt.key === 'Esc') {
-        evt.preventDefault();
-        replaceItemFormToItem();
-        document.removeEventListener('keydown', onEscKeyDown);
-      }
-    };
-
-    eventComponent
-      .setEditClickHandler(() => {
-        replaceItemToItemForm();
-        document.addEventListener('keydown', onEscKeyDown);
-      });
-
-    eventEditComponent
-      .setFormSubmitHandler(() => {
-        replaceItemFormToItem();
-        document.removeEventListener('keydown', onEscKeyDown);
-      });
-
-    render(this._eventsListComponent, eventComponent, RenderPosition.BEFORE_END);
+    const eventPresenter = new EventPresenter(this._eventsListComponent);
+    eventPresenter.init(event);
   }
 
   _renderEventItems() {
